requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.11
python-dateutil==2.8.2
lxml==5.1.0
